zaj luqibase
