+++ src/main/resources/application-no-liquibase.properties
@@ -1,0 +1,3 @@
+spring.datasource.url=jdbc:h2:file:~/library-api-no-liquibase
+spring.jpa.hibernate.ddl-auto=create-drop
+spring.liquibase.enabled=false
