+++ src/main/resources/application.properties
@@ -12,6 +12,7 @@ spring.h2.console.enabled=true
 spring.jpa.hibernate.ddl-auto=validate
 spring.jpa.open-in-view=false
 spring.liquibase.change-log=classpath:liquibase-master.xml
+#spring.liquibase.enabled=false
 
 spring.servlet.multipart.max-file-size=2GB
 spring.servlet.multipart.max-request-size=2GB
