+++ src/main/resources/changes/create-tables.xml
@@ -6,7 +6,7 @@
 
     <changeSet id="20240513-krzysiek-001" author="krzysiek">
         <sql>
-            create sequence person_seq start with 1 increment by 100
+            create sequence person_seq start with 100 increment by 100
         </sql>
     </changeSet>
 
@@ -14,7 +14,6 @@
         <sql>
             create table author (id integer generated by default as identity, birth_year integer, death_year integer,
             name varchar(255), surname varchar(255), primary key (id))
-
         </sql>
     </changeSet>
 
@@ -22,6 +21,43 @@
         <sql>
             create table book (id integer generated by default as identity, available boolean not null, category
             varchar(255), book_title varchar(500), author_id integer, primary key (id))
+        </sql>
+    </changeSet>
+
+    <changeSet id="20240513-krzysiek-004" author="krzysiek">
+        <sql>
+            create table car (id integer generated by default as identity, brand_name varchar(500), fuel_type
+            varchar(255), model varchar(255), garage_id integer, primary key (id))
+        </sql>
+    </changeSet>
+
+    <changeSet id="20240513-krzysiek-005" author="krzysiek">
+        <sql>
+            create table garage (id integer generated by default as identity, address varchar(255), lpg_allowed boolean,
+            places integer not null, primary key (id))
+        </sql>
+    </changeSet>
+    <changeSet id="20240513-krzysiek-006" author="krzysiek">
+        <sql>
+            create table import_status (id integer generated by default as identity, failed_reason varchar(255),
+            file_name varchar(255), finish_date timestamp(6), processed integer not null, start_date timestamp(6),
+            status smallint, submit_date timestamp(6), primary key (id))
+        </sql>
+    </changeSet>
+    <changeSet id="20240513-krzysiek-007" author="krzysiek">
+        <sql>
+            create table person (dtype varchar(31) not null, id integer not null, age integer not null, name
+            varchar(255), pesel varchar(255), position varchar(255), salary integer, type varchar(255), std_group
+            varchar(255), scholarship integer, primary key (id))
+        </sql>
+    </changeSet>
+    <changeSet id="20240513-krzysiek-008" author="krzysiek">
+        <sql>
+            alter table if exists person add constraint UK_PERSON_PESEL unique (pesel);
+            alter table if exists book add constraint FK_BOOK_AUTHOR foreign key (author_id) references
+            author;
+            alter table if exists car add constraint FK_CAR_GARAGE foreign key (garage_id) references
+            garage;
         </sql>
     </changeSet>
